requirements.txt
boto3>=1.28.0
pydantic>=2.5.0
openapi-spec-validator>=0.7.1
PyYAML>=6.0.1
Jinja2>=3.1.2
click>=8.1.7
black>=23.12.0
mypy>=1.7.1
pytest>=7.4.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
